Made the Call To Action a CSS class.
There may be more than one in the document, and it's not valid to have
two elements with the same ID so it should be a class.

It was also corrected where it was used, and a broken link was fixed
on the 404 page.

Fixed the remaining invalid HTML.

(Apart from one instance of a h4 being inside a paragraph, but that
doesn't really mattter.)

Fixed the last HTML validation error.

Removed frameborder - it's no longer valid HTML5.